Major refactoring, split Makefile.gather from Makefile.download
In order to avoid the .download step when the manifest indicates
that no component has changed in the slightest, we need to pre-
process the manifest (remains Makefile.gather(intel)) resulting
in the list of packages, now with further annotations about the
package or repo rev required. Slurping down the packages now
becomes the task of Makefile.download, to be followed by the
unchanged Makefile.preconfig phase.

Anyone with a toolchain based on this repo is advised to see
Phase 1 of the README and add the extra Makefile.download step
following Makefile.gather.

This commit is somewhat incomplete, directory names as targets
appear to trigger without pause or remorse. The remaining
TODO: ensure we don't redownload packages on subsequent tries
to invoke Makefile.download.